restructure code and use buffer instead of file for capturing stdout
