--- dgraph/cmd/live/wrong-filename/docker-compose.yml
@@ -1,44 +1,0 @@
-version: "3.5"
-services:
-  zero1:
-    image: dgraph/dgraph:latest
-    container_name: zero1
-    working_dir: /data/zero1
-    ports:
-      - 5180:5180
-      - 6180:6180
-    labels:
-      cluster: test
-      service: zero
-    volumes:
-      - type: bind
-        source: $GOPATH/bin
-        target: /gobin
-        read_only: true
-    command: /gobin/dgraph zero -o 100 --my=zero1:5180 --idx 1 --logtostderr -v=2 --bindall --expose_trace --profile_mode block --block_rate 10
-
-  alpha1:
-    image: dgraph/dgraph:latest
-    container_name: alpha1
-    working_dir: /data/alpha1
-    volumes:
-      - type: bind
-        source: $GOPATH/bin
-        target: /gobin
-        read_only: true
-      - type: bind
-        source: ../../../../ee/acl/hmac-secret
-        target: /dgraph-acl/hmac-secret
-        read_only: true
-      - type: bind
-        source: ../../../../ee/enc/test-fixtures/enc-key
-        target: /dgraph-enc/enc-key
-        read_only: true
-    ports:
-      - 8180:8180
-      - 9180:9180
-    labels:
-      cluster: test
-      service: alpha
-    command: /gobin/dgraph alpha --encryption_key_file "/dgraph-enc/enc-key" --my=alpha1:7180 --lru_mb=1024 --zero=zero1:5180 -o 100 --expose_trace --trace 1.0 --profile_mode block --block_rate 10 --logtostderr -v=2 --whitelist 10.0.0.0/8,172.16.0.0/12,192.168.0.0/16  --acl_secret_file /dgraph-acl/hmac-secret --acl_access_ttl 3s --acl_cache_ttl 5s
-
