--- dgraph/cmd/live/wrong-filename/filename_test.go
@@ -1,93 +1,0 @@
-/*
- * Copyright 2017-2018 Dgraph Labs, Inc. and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package main
-
-import (
-	"context"
-	"io/ioutil"
-	"log"
-	"os"
-	"path"
-	"runtime"
-	"strings"
-	"testing"
-
-	"github.com/dgraph-io/dgo/v200"
-	"github.com/dgraph-io/dgo/v200/protos/api"
-	"github.com/stretchr/testify/require"
-
-	"github.com/dgraph-io/dgraph/testutil"
-	"github.com/dgraph-io/dgraph/x"
-)
-
-var alphaService = testutil.SockAddr
-var zeroService = testutil.SockAddrZero
-
-var (
-	testDataDir string
-	dg          *dgo.Dgraph
-)
-
-const (
-	alphaName       = "alpha1"
-	alphaExportPath = alphaName + ":/data/" + alphaName + "/export"
-	localExportPath = "./export_copy"
-)
-
-func extractFileName(output string) string {
-	lines := strings.Split(output, "\n")
-	line := lines[7]
-	filename := strings.Split(strings.Split(line, testDataDir)[1], ":")[0]
-	return filename[1 : len(filename)-1]
-}
-
-func TestLiveLoadFileName(t *testing.T) {
-	testutil.DropAll(t, dg)
-
-	pipeline := [][]string{
-		{testutil.DgraphBinaryPath(), "live",
-			"--files", testDataDir + "/a.rdf," + testDataDir + "/b.rdf",
-			"--alpha", alphaService, "--zero", zeroService, "-u", "groot", "-p", "password"},
-	}
-
-	out, err := testutil.Pipeline(pipeline)
-	name := extractFileName(string(out))
-	require.Error(t, err, "live loading of rdf file did not exit with error")
-	require.Equal(t, name, "b.rdf", "incorrect name for file")
-}
-
-func TestMain(m *testing.M) {
-	_, thisFile, _, _ := runtime.Caller(0)
-	testDataDir = path.Dir(thisFile)
-
-	var err error
-	dg, err = testutil.DgraphClientWithGroot(testutil.SockAddr)
-	if err != nil {
-		log.Fatalf("Error while getting a dgraph client: %v", err)
-	}
-	x.Check(dg.Alter(
-		context.Background(), &api.Operation{DropAll: true}))
-
-	// Try to create any files in a dedicated temp directory that gets cleaned up
-	// instead of all over /tmp or the working directory.
-	tmpDir, err := ioutil.TempDir("", "test.tmp-")
-	x.Check(err)
-	os.Chdir(tmpDir)
-	defer os.RemoveAll(tmpDir)
-
-	os.Exit(m.Run())
-}
